Updated CoffeeMaker functionality

+++ src/CoffeeMaker/Drink.cs
@@ -12,9 +12,9 @@ namespace CoffeeMaker
         private string[] _ingredientNames = { "water", "cream", "coffee", "sugar" };
         private string[] _units = { "oz", "oz", "g", "g" };
         private int[] _ingredientsUsed = { 0, 0, 0, 0 };
-        private double _cost = 0;
+        private decimal _cost = 0;
 
-        public Drink(string name, double cost, int water, int cream, int coffee, int sugar)
+        public Drink(string name, decimal cost, int water, int cream, int coffee, int sugar)
         {
             _name = name;
             _ingredientsUsed[0] = water;
